[Step4] refactor : 코드 리뷰 반영 - view 관심사 이동
