+++ src/main/kotlin/blackjack/view/ResultView.kt
@@ -30,11 +30,11 @@ class ResultView {
 
     private fun printCardsByPlayer(player: Player, withScore: Boolean) {
         if (withScore) {
-            println("${player.name}카드: ${player.receivedCards.getCardDescription()} - 결과: ${player.score}")
+            println("${player.name}카드: ${CardView.getCardDescription(player.receivedCards)} - 결과: ${player.score}")
             return
         }
 
-        println("${player.name}카드: ${player.receivedCards.getCardDescription()}")
+        println("${player.name}카드: ${CardView.getCardDescription(player.receivedCards)}")
     }
 
     private fun getDealerResult(dealer: Dealer, player: List<Player>): String {
